refactor(msteams): isolate inbound dispatch lifecycle
