+++ extensions/msteams/src/monitor-handler/message-handler.ts
@@ -1,37 +1,27 @@
 // Msteams plugin module implements message handler behavior.
 import {
-  buildChannelInboundEventContext,
-  createChannelInboundEnvelopeBuilder,
   formatMediaPlaceholderText,
   resolveInboundMentionDecision,
-  resolveInboundSupplementalSenderAllowed,
-  toInboundMediaFactsWithMetadata,
-} from "openclaw/plugin-sdk/channel-inbound";
-import {
-  hasFinalInboundReplyDispatch,
-  resolveInboundReplyDispatchCounts,
 } from "openclaw/plugin-sdk/channel-inbound";
 import { fanInChannelIngressLifecycles } from "openclaw/plugin-sdk/channel-ingress-runtime";
-import { bindIngressLifecycleToReplyOptions } from "openclaw/plugin-sdk/channel-outbound";
 import { resolveChannelContextVisibilityMode } from "openclaw/plugin-sdk/context-visibility-runtime";
 import {
   DEFAULT_GROUP_HISTORY_LIMIT,
   createChannelHistoryWindow,
   type HistoryEntry,
 } from "openclaw/plugin-sdk/reply-history";
-import { sliceUtf16Safe, truncateUtf16Safe } from "openclaw/plugin-sdk/text-utility-runtime";
+import { truncateUtf16Safe } from "openclaw/plugin-sdk/text-utility-runtime";
 import { formatUnknownError } from "../errors.js";
 import { normalizeMSTeamsConversationId, parseMSTeamsActivityTimestamp } from "../inbound.js";
 import type { MSTeamsMessageHandlerDeps } from "../monitor-handler.types.js";
 import type { MSTeamsIngressLifecycle } from "../msteams-ingress.js";
-import { resolveMSTeamsAllowlistMatch, resolveMSTeamsReplyPolicy } from "../policy.js";
+import { resolveMSTeamsReplyPolicy } from "../policy.js";
 import { extractMSTeamsPollVote } from "../polls.js";
-import { createMSTeamsReplyDispatcher } from "../reply-dispatcher.js";
 import { getMSTeamsRuntime } from "../runtime.js";
 import type { MSTeamsTurnContext } from "../sdk-types.js";
-import { recordMSTeamsSentMessage } from "../sent-message-cache.js";
 import { admitMSTeamsMessage } from "./access.js";
 import { prepareMSTeamsInboundContent } from "./inbound-content.js";
+import { dispatchMSTeamsInboundTurn } from "./inbound-dispatch.js";
 import {
   assembleMSTeamsInboundFacts,
   prepareMSTeamsDebounceEntry,
@@ -95,7 +85,6 @@ export function createMSTeamsMessageHandler(deps: MSTeamsMessageHandlerDeps) {
       conversationType,
       isChannel,
       teamId,
-      graphChannelId,
       conversationRef,
     } = facts;
     const historyBody = [text, formatMediaPlaceholderText(advertisedMedia)]
@@ -173,13 +162,6 @@ export function createMSTeamsMessageHandler(deps: MSTeamsMessageHandlerDeps) {
       return;
     }
 
-    const teamsFrom = isDirectMessage
-      ? `msteams:${senderId}`
-      : isChannel
-        ? `msteams:channel:${conversationId}`
-        : `msteams:group:${conversationId}`;
-    const teamsTo = isDirectMessage ? `user:${senderId}` : `conversation:${conversationId}`;
-
     const threadRouting = prepareMSTeamsThreadRouting({
       cfg,
       context,
@@ -193,7 +175,6 @@ export function createMSTeamsMessageHandler(deps: MSTeamsMessageHandlerDeps) {
     });
     const { route, deadline: preprocessingDeadline } = threadRouting;
 
-    const preview = sliceUtf16Safe(rawBody.replace(/\s+/g, " "), 0, 160);
     const inboundLabel = isDirectMessage
       ? `Teams DM from ${senderName}`
       : `Teams message in ${conversationType} from ${senderName}`;
@@ -274,264 +255,44 @@ export function createMSTeamsMessageHandler(deps: MSTeamsMessageHandlerDeps) {
     if (!content) {
       return;
     }
-    const { agentBody, inboundMedia } = content;
     enqueuePrimaryMessageSystemEvent();
 
-    const { teamAadGroupId, quoteBodyFull, quoteSenderId, quoteSenderName, threadContext } =
-      await resolveMSTeamsThreadContext({
-        routing: threadRouting,
-        context,
-        tokenProvider,
-        quoteInfo,
-        isDirectMessage,
-        isChannel,
-        conversationId,
-        contextVisibilityMode,
-        groupPolicy,
-        effectiveGroupAllowFrom,
-        allowNameMatching,
-        log,
-      });
-
-    const envelopeFrom = isDirectMessage ? senderName : conversationType;
-    const buildEnvelope = createChannelInboundEnvelopeBuilder({ cfg, route });
-    const body = buildEnvelope({
-      channel: "Teams",
-      from: envelopeFrom,
-      timestamp,
-      body: agentBody,
-    });
-    let combinedBody = body;
-    const isRoomish = !isDirectMessage;
-    const historyKey = isRoomish ? conversationId : undefined;
-    if (isRoomish && historyKey) {
-      const channelHistory = createChannelHistoryWindow({ historyMap: conversationHistories });
-      combinedBody = channelHistory.buildPendingContext({
-        historyKey,
-        limit: historyLimit,
-        currentMessage: combinedBody,
-        formatEntry: (entry) =>
-          buildEnvelope({
-            channel: "Teams",
-            from: conversationType,
-            timestamp: entry.timestamp,
-            previousTimestamp: null,
-            body: `${entry.sender}: ${entry.body}${entry.messageId ? ` [id:${entry.messageId}]` : ""}`,
-          }),
-      });
-    }
-
-    const inboundHistory =
-      isRoomish && historyKey && historyLimit > 0
-        ? createChannelHistoryWindow({ historyMap: conversationHistories }).buildInboundHistory({
-            historyKey,
-            limit: historyLimit,
-          })
-        : undefined;
-    const commandBody = text.trim();
-    const quoteSenderAllowed =
-      quoteInfo && quoteInfo.sender
-        ? resolveInboundSupplementalSenderAllowed({
-            isGroup: !isDirectMessage,
-            groupPolicy,
-            allowFrom: effectiveGroupAllowFrom,
-            isSenderAllowed: (allowFrom) =>
-              resolveMSTeamsAllowlistMatch({
-                allowFrom,
-                senderId: quoteSenderId ?? "",
-                senderName: quoteSenderName,
-                allowNameMatching,
-              }).allowed,
-          })
-        : true;
-    // Prepend thread history to the agent body so the agent has full thread context.
-    const bodyForAgent = threadContext
-      ? `[Thread history]\n${threadContext}\n[/Thread history]\n\n${agentBody}`
-      : agentBody;
-
-    // For Teams *channel* messages (not group chats / DMs), preserve the
-    // `aadGroupId/channelId` pair on NativeChannelId so downstream action handlers
-    // can route through `/teams/{aadGroupId}/channels/{channelId}` via Graph API.
-    // The bare conversation id (`19:...@thread.tacv2`) is insufficient on its
-    // own because channel Graph endpoints require the owning team id too.
-    const nativeChannelId =
-      isChannel && teamAadGroupId ? `${teamAadGroupId}/${graphChannelId}` : undefined;
-    const ctxPayload = buildChannelInboundEventContext({
-      channel: "msteams",
-      contextVisibility: contextVisibilityMode,
-      supplemental: {
-        quote: quoteInfo
-          ? {
-              id: quoteInfo.id ?? activity.replyToId ?? undefined,
-              body: quoteBodyFull ?? quoteInfo.body,
-              sender: quoteInfo.sender,
-              senderAllowed: quoteSenderAllowed,
-              isQuote: true,
-            }
-          : undefined,
-      },
-      media: await toInboundMediaFactsWithMetadata(inboundMedia),
-      messageId: activity.id,
-      timestamp: timestamp?.getTime() ?? Date.now(),
-      from: teamsFrom,
-      sender: {
-        id: senderId,
-        name: senderName,
-      },
-      conversation: {
-        kind: isDirectMessage ? "direct" : isChannel ? "channel" : "group",
-        id: conversationId,
-        label: envelopeFrom,
-        spaceId: teamId,
-        nativeChannelId,
-      },
-      route: {
-        agentId: route.agentId,
-        dmScope: route.dmScope,
-        accountId: route.accountId,
-        routeSessionKey: route.sessionKey,
-      },
-      reply: {
-        to: teamsTo,
-        replyToId: activity.replyToId ?? undefined,
-        nativeChannelId,
-      },
-      message: {
-        body: combinedBody,
-        bodyForAgent,
-        inboundHistory,
-        rawBody,
-        commandBody,
-      },
-      sessionTranscript: { historyLimit: isRoomish ? historyLimit : 0 },
-      access: {
-        mentions: {
-          canDetectMention: !isDirectMessage,
-          wasMentioned: isDirectMessage || mentionDecision.effectiveWasMentioned,
-        },
-        commands: {
-          authorized: commandAuthorized === true,
-        },
-      },
-      extra: {
-        GroupSubject: !isDirectMessage ? conversationType : undefined,
-        ReplyToIsQuote: quoteInfo ? true : undefined,
-      },
-    });
-
-    logVerboseMessage(`msteams inbound: from=${ctxPayload.From} preview="${preview}"`);
-
-    const sharePointSiteId = msteamsCfg?.sharePointSiteId;
-    const { dispatcherOptions, delivery, replyOptions } = createMSTeamsReplyDispatcher({
-      cfg,
-      agentId: route.agentId,
-      sessionKey: route.sessionKey,
-      accountId: route.accountId,
-      runtime,
-      log,
-      app,
-      appId,
-      conversationRef,
+    const thread = await resolveMSTeamsThreadContext({
+      routing: threadRouting,
       context,
-      replyStyle,
-      textLimit,
-      onSentMessageIds: (ids) => {
-        for (const id of ids) {
-          recordMSTeamsSentMessage(conversationId, id);
-        }
-      },
       tokenProvider,
-      sharePointSiteId,
+      quoteInfo,
+      isDirectMessage,
+      isChannel,
+      conversationId,
+      contextVisibilityMode,
+      groupPolicy,
+      effectiveGroupAllowFrom,
+      allowNameMatching,
+      log,
     });
 
-    // Use Teams clientInfo timezone if no explicit userTimezone is configured.
-    // This ensures the agent knows the sender's timezone for time-aware responses
-    // and proactive sends within the same session.
-    const activityClientInfo = activity.entities?.find((e) => e.type === "clientInfo") as
-      | { timezone?: string }
-      | undefined;
-    const senderTimezone = activityClientInfo?.timezone || conversationRef.timezone;
-    const turnConfig =
-      senderTimezone && !cfg.agents?.defaults?.userTimezone
-        ? {
-            ...cfg,
-            agents: {
-              ...cfg.agents,
-              defaults: { ...cfg.agents?.defaults, userTimezone: senderTimezone },
-            },
-          }
-        : cfg;
-    log.info("dispatching to agent", { sessionKey: route.sessionKey });
-    try {
-      const turnResult = await core.channel.inbound.run({
-        channel: "msteams",
-        accountId: route.accountId,
-        raw: context,
-        adapter: {
-          ingest: () => ({
-            id: activity.id ?? `${teamsFrom}:${Date.now()}`,
-            timestamp: timestamp?.getTime(),
-            rawText: rawBody,
-            textForAgent: bodyForAgent,
-            textForCommands: commandBody,
-            raw: activity,
-          }),
-          resolveTurn: () => ({
-            cfg: turnConfig,
-            channel: "msteams",
-            accountId: route.accountId,
-            route: { agentId: route.agentId, sessionKey: route.sessionKey },
-            ctxPayload,
-            record: {
-              onRecordError: (err) => {
-                logVerboseMessage(
-                  `msteams: failed updating session meta: ${formatUnknownError(err)}`,
-                );
-              },
-            },
-            history: {
-              isGroup: isRoomish,
-              historyKey,
-              historyMap: conversationHistories,
-              limit: historyLimit,
-            },
-            dispatcherOptions,
-            delivery,
-            replyOptions: {
-              ...replyOptions,
-              ...(params.turnAdoptionLifecycle
-                ? bindIngressLifecycleToReplyOptions(params.turnAdoptionLifecycle)
-                : {}),
-            },
-          }),
-        },
-      });
-      const dispatchResult = turnResult.dispatched ? turnResult.dispatchResult : undefined;
-      const queuedFinal = dispatchResult?.queuedFinal ?? false;
-      const counts = resolveInboundReplyDispatchCounts(dispatchResult);
-      const hasFinalResponse = hasFinalInboundReplyDispatch(dispatchResult);
-
-      log.info("dispatch complete", { queuedFinal, counts });
-
-      if (!hasFinalResponse) {
-        return;
-      }
-      const finalCount = counts.final;
-      logVerboseMessage(
-        `msteams: delivered ${finalCount} reply${finalCount === 1 ? "" : "ies"} to ${teamsTo}`,
-      );
-    } catch (err) {
-      log.error("dispatch failed", { error: formatUnknownError(err) });
-      runtime.error(`msteams dispatch failed: ${formatUnknownError(err)}`);
-      if (params.turnAdoptionLifecycle) {
-        throw err;
-      }
-      try {
-        await context.sendActivity("⚠️ Something went wrong. Please try again.");
-      } catch {
-        // Best effort.
-      }
-    }
+    await dispatchMSTeamsInboundTurn({
+      cfg,
+      runtime,
+      appId,
+      app,
+      tokenProvider,
+      textLimit,
+      log,
+      logVerboseMessage,
+      facts,
+      admission,
+      content,
+      routing: threadRouting,
+      thread,
+      replyStyle,
+      timestamp,
+      contextVisibilityMode,
+      mentionWasEffective: mentionDecision.effectiveWasMentioned,
+      conversationHistories,
+      historyLimit,
+    });
   };
 
   const inboundDebouncer = core.channel.debounce.createInboundDebouncer<MSTeamsDebounceEntry>({
@@ -623,4 +384,3 @@ export function createMSTeamsMessageHandler(deps: MSTeamsMessageHandlerDeps) {
     return undefined;
   };
 }
-/* oxlint-disable max-lines -- TODO: split this grandfathered oversized file. */
